+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## unreleased
+
+- Save whether the automatic mode is enabled or not
+- On devices without light sensor the app now starts and runs in manual mode only.
+
 ## 1.0.1
 
-* Show message when no update is available.
-* Show an error message if the device has no light sensor.
+- Show message when no update is available.
+- Show an error message if the device has no light sensor.
 
 ## 1.0.0